requirements.txt
bcrypt==3.2.2
boto3==1.24.39
botocore==1.27.39
cffi==1.15.1
cryptography==37.0.4
jmespath==1.0.1
paramiko==2.11.0
psycopg2-binary==2.9.3
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.11
